change everything to u64s

+++ contracts/src/types.cairo
@@ -1,8 +1,8 @@
 
 #[derive(Copy, Drop, Serde, Introspect)]
 struct Vec2 {
-    x: u32,
-    y: u32,
+    x: u64,
+    y: u64,
 } 
 
 #[derive(Copy, Drop, Serde, Introspect)]
